# Avoid promise wrapping for middleware next request handlers and add documentation for consuming response from next request handler